업그레이드 float->int 변경

+++ TimeTraveler/Assets/Script/Data/SaveLoadManager.cs
@@ -9,9 +9,9 @@ public class GameData
 {
     public int high_score;
     public int coin;
-    public float upgrade_hp;
+    public int upgrade_hp;
     public int upgrade_jelly;
-    public float upgrade_energy;
+    public int upgrade_energy;
     public int heart;
     public string exit_time;
     public int buff;
@@ -80,7 +80,7 @@ public class SaveLoadManager : MonoBehaviour
         savePath = Path.Combine(Application.persistentDataPath, "PlayerData.json");
     }
 
-    public void SaveGameData(int score, int coin, float upgrade_hp, int upgrade_jelly, float upgrade_energy)
+    public void SaveGameData(int score, int coin, int upgrade_hp, int upgrade_jelly, int upgrade_energy)
     {
         _gameData.high_score = score;
         _gameData.coin = coin;
@@ -118,8 +118,8 @@ public class SaveLoadManager : MonoBehaviour
     public void UpgradeHP() { GameData.upgrade_hp++; } //호출 시 upgrade_hp 1 증가
     public void UpgradeEnergy() { GameData.upgrade_energy++; } //호출 시 upgrade_energy 1 증가
     public void UpgradeJelly() { GameData.upgrade_jelly++; } //호출 시 upgrade_jelly 1 증가
-    public float GetUpgradeHP() { return GameData.upgrade_hp; } //upgrade_hp 반환
-    public float GetUpgradeEnergy() { return GameData.upgrade_energy; } //upgrade_energy 반환
+    public int GetUpgradeHP() { return GameData.upgrade_hp; } //upgrade_hp 반환
+    public int GetUpgradeEnergy() { return GameData.upgrade_energy; } //upgrade_energy 반환
     public int GetUpgradeJelly() { return GameData.upgrade_jelly; } //upgrade_jelly 반환
     public void PlusCoin(int coin) { GameData.coin += coin; } //Coin 더하기
     public void MinusCoin(int coin) { GameData.coin -= coin; } //Coin 빼기
